// src/app/core/trust-and-safety/report/enums/item-report-reasons.enum.ts
export enum ITEM_REPORT_REASON_VALUE {
  PEOPLE_ANIMALS = 'people_or_animals',
  JOKE = 'joke',
  FAKE = 'fake_item',
  EXPLICIT = 'explicit_content',
  WRONG_IMAGE = 'wrong_image',
  FOOD = 'food',
  DRUGS = 'drugs_or_medicines',
  DUPLICATED = 'duplicate_product',
  FORBIDDEN = 'forbidden_product_service',
  RESALE = 'resale',
  AD_SPAM = 'ad_or_spam',
}